Fix USB type A connector
The USB type A connector footprint is currently not symmetrical (this
can be seen on the 3D view). It looks like that some pins have been
aligned to a mil grid, while the distance between pins on a USB
connector should be aligned on a metric grid.

Fix this by using setting the distance between been pins 2 and 3 to 2mm
and the distance between pins 1 and 2 and between pins 3 and 4 to 2.5mm.
This matches the distance between signals on a female A connector. Also
slightly change the dimensions of the connector to make it fully
symmetrical.

The changes are small enough so that it does not affect the 3D view (it
even looks better now).